refactor(qn): split the constants, frame builders and matching out of the class
qn-scale/index.ts drops from 1636 lines to 1144. What left:

  constants.ts   the UUIDs, report bytes, thresholds and the decode narrative
                 attached to them
  frames.ts      buildMeasurementTrigger, buildA2Frame, buildTimeSync
  matching.ts    matches() and warnOnOneByoneShape()
  broadcast.ts   parseBroadcast()

The split is by purity, not by the seven-module table in the issue. Every one of
the four moved units reads zero fields of the adapter instance, verified rather
than assumed; matches() alone is 95 lines with not a single `this.` in it.

What stayed, and why it is not an oversight: parseNotification touches 16
distinct members, onConnected touches 26, and the four handshake steps between
them touch fifteen fields plus three methods. Moving those means either passing
the instance around or dropping `private` from twenty-five fields so five
modules can reach into the class's state. The hazard this issue names is "the
next change to one dialect quietly moves behaviour in the other", and widening
that surface makes it more likely, not less.

Verified as a pure move the same way as 39c74ae: comparing the file before
against the concatenation of the modules after, ignoring imports, every
difference is a `const` gaining `export`, a moved function gaining `export`, one
of the three delegate lines, or a new module header. No logic line changed.

The 1950-line test file passes with no assertion edit, which is the whole
safety net for this and the reason it was done in this order.

Two things the move made safe rather than carried over:
  - buildA2Frame now says in the file not to merge it with the ready-time A2.
    They are the identical six-byte shape and buildA2Frame((0x32 << 8) | age)
    would emit the same bytes, but the scale reads their payloads differently.
  - constants.ts leads with a note that most of it is evidence, not code, and
    that PROTOCOLS.md cites it.

index.ts is still well above the ~400 the issue asks for. That is the honest
outcome of the paragraph above rather than a step left undone.

Refs #367
